feat: 팀 목록 조회시 정렬 및 필터링 기능 제거
- TeamSearchCondition을 관리 대상에서 제외한다.
- 이전 버전 클라이언트 의존성으로 인해 내부적으로 사용되지 않는 TeamSearchCondition 입력을 받는다.
