--- src/main/java/com/ssh/dartserver/domain/team/domain/TeamSearchCondition.java
@@ -1,10 +1,0 @@
-package com.ssh.dartserver.domain.team.domain;
-
-import lombok.Data;
-@Data
-public class TeamSearchCondition {
-    private OrderMethod order;
-    private Long regionId;
-    private Boolean verifiedStudent;
-    private Boolean hasProfileImage;
-}
