+++ src/main/java/com/ssh/dartserver/domain/team/presentation/TeamController.java
@@ -2,7 +2,7 @@ package com.ssh.dartserver.domain.team.presentation;
 
 import com.ssh.dartserver.domain.team.application.MyTeamService;
 import com.ssh.dartserver.domain.team.application.TeamService;
-import com.ssh.dartserver.domain.team.domain.TeamSearchCondition;
+import com.ssh.dartserver.domain.team.presentation.request.TeamSearchCondition;
 import com.ssh.dartserver.domain.team.presentation.request.TeamRequest;
 import com.ssh.dartserver.domain.team.presentation.response.BlindDateTeamDetailResponse;
 import com.ssh.dartserver.domain.team.presentation.response.BlindDateTeamResponse;
@@ -56,10 +56,11 @@ public class TeamController {
         return ResponseEntity.ok(teamService.readTeam(principal.getUser(), id));
     }
 
+    // 이전 버전 클라이언트 의존성으로 인해 내부적으로 사용되지 않는 TeamSearchCondition 입력을 받는다.
     @GetMapping
-    public ResponseEntity<Page<BlindDateTeamResponse>> listTeam(Authentication authentication, TeamSearchCondition condition, Pageable pageable) {
+    public ResponseEntity<Page<BlindDateTeamResponse>> listTeam(Authentication authentication, @Deprecated TeamSearchCondition condition, Pageable pageable) {
         PrincipalDetails principal = (PrincipalDetails) authentication.getPrincipal();
-        return ResponseEntity.ok(teamService.listVisibleTeam(principal.getUser(), condition, pageable));
+        return ResponseEntity.ok(teamService.listVisibleTeam(principal.getUser(), pageable));
     }
 
 }
